fix: gate git secret create on feature flag; drop unused GitSecretField
